+++ Libraries/plist/Sources/Format/JSONWriter.cpp
@@ -184,7 +184,7 @@ handleDictionary(Dictionary const *dictionary, bool root)
 
     _lastKey = false;
 
-    for (int i = 0; i < dictionary->count(); ++i) {
+    for (size_t i = 0; i < dictionary->count(); ++i) {
         /* Write ',' if not first entry. */
         if (i != 0) {
             if (!writeString(",\n", false)) {
@@ -234,7 +234,7 @@ handleArray(Array const *array, bool root)
 
     _indent++;
 
-    for (int i = 0; i < array->count(); ++i) {
+    for (size_t i = 0; i < array->count(); ++i) {
         /* Write ',' if not first entry. */
         if (i != 0) {
             if (!writeString(",\n", false)) {
